🚧 WIP on inline TipTapEditor
