+++ gui/src/editorInset/EditorInset.tsx
@@ -1,31 +1,56 @@
-import { useSelector } from "react-redux";
+import { useEffect, useRef } from "react";
+import { useDispatch, useSelector } from "react-redux";
 import styled from "styled-components";
+import { defaultBorderRadius } from "../components";
 import TipTapEditor from "../components/mainInput/TipTapEditor";
+import useSetup from "../hooks/useSetup";
 import { selectSlashCommands } from "../redux/selectors";
 import { RootState } from "../redux/store";
+import { ideRequest } from "../util/ide";
 
 const EditorInsetDiv = styled.div`
   max-width: 500px;
+  position: relative;
+  display: flex;
+  border-radius: ${defaultBorderRadius};
+  // box-shadow: 0 0 8px 0 rgba(0, 0, 0, 0.4);
 `;
 
 function EditorInset() {
+  const dispatch = useDispatch();
   const availableSlashCommands = useSelector(selectSlashCommands);
   const availableContextProviders = useSelector(
     (store: RootState) => store.state.config.contextProviders,
   );
 
+  useSetup(dispatch);
+
+  const elementRef = useRef(null);
+
+  useEffect(() => {
+    if (!elementRef.current) return;
+    const resizeObserver = new ResizeObserver(() => {
+      if (!elementRef.current) return;
+
+      console.log("Height: ", elementRef.current.clientHeight);
+      ideRequest("jetbrains/editorInsetHeight", {
+        height: elementRef.current.clientHeight,
+      });
+    });
+    resizeObserver.observe(elementRef.current);
+    return () => resizeObserver.disconnect();
+  }, []);
+
   return (
-    <EditorInsetDiv>
-      <div className="flex px-2 relative">
-        <TipTapEditor
-          availableContextProviders={availableContextProviders}
-          availableSlashCommands={availableSlashCommands}
-          isMainInput={true}
-          onEnter={(e) => {
-            console.log("Enter: ", e);
-          }}
-        ></TipTapEditor>
-      </div>
+    <EditorInsetDiv ref={elementRef}>
+      <TipTapEditor
+        availableContextProviders={availableContextProviders}
+        availableSlashCommands={availableSlashCommands}
+        isMainInput={true}
+        onEnter={(e) => {
+          console.log("Enter: ", e);
+        }}
+      ></TipTapEditor>
     </EditorInsetDiv>
   );
 }
